refactor(playground): migrate e2e fixtures to folder-routed daemons
`playground/opensidian-e2e/` and `playground/tab-manager-e2e/` were
standalone Bun scripts invoked via `bun run epicenter.config.ts`. They
now boot through the canonical loader: each playground has its own
`workspaces/<route>/daemon.ts` that `epicenter daemon up -C <playground>`
discovers and opens.

The opensidian playground's giant standalone module collapses to one
`defineDaemonWorkspace({ open })` that uses the host-derived
`attachEncryption` and `openWebSocket`. Its renamed test
`workspace.test.ts` exercises the same `opensidianTables` schema
end-to-end without needing the CLI binary.

The tab-manager playground keeps its markdown materializer chain;
auth handling and replicaId now come from the daemon ctx instead of
`createMachineAuthClient` inside the module.

`push-from-markdown.ts` references the new daemon type
(`OpensidianPlaygroundRuntime`) and the README points readers at
`epicenter daemon up` instead of `bun run epicenter.config.ts`.
